Move pool routing types under the internal package
The pool registry, execution-target router, and claim planner were
created under ri.core but match the rest of the post-hardening
framework internals. Move them under ri.core.internal so the layout
stays consistent and update the four callers outside that package
(RatchetProducer, JobSubmissionService, RetryBufferDrainer, and
SubmissionGateChecker) to import them through the new path.

Also patches two TCK helper builders (the JPA recurring-claim
concurrency contract and the Mongo recurring-job contract test) that
still constructed RecurringJobDefinition with the pre-executionTarget
arity.
