+++ ratchet-store-mongodb/src/test/java/run/ratchet/store/mongodb/MongoRecurringJobStoreContractTest.java
@@ -120,6 +120,7 @@ class MongoRecurringJobStoreContractTest extends AbstractRecurringJobStoreContra
         null,
         null,
         null,
+        null,
         Instant.now(),
         null);
   }
