+++ ratchet-tck/store/src/main/java/run/ratchet/tck/store/AbstractJpaRecurringClaimConcurrencyContract.java
@@ -160,6 +160,7 @@ public abstract class AbstractJpaRecurringClaimConcurrencyContract {
         null,
         null,
         null,
+        null,
         Instant.now(),
         null);
   }
