+++ ratchet/src/main/java/run/ratchet/ri/cdi/RatchetProducer.java
@@ -20,7 +20,6 @@ import run.ratchet.api.ExecutorTargets;
 import run.ratchet.api.RatchetOptions;
 import run.ratchet.ri.core.DrainController;
 import run.ratchet.ri.core.PollerScheduler;
-import run.ratchet.ri.core.PoolRegistry;
 import run.ratchet.ri.core.ResourcePermitService;
 import run.ratchet.ri.core.internal.DefaultNodeIdentityProvider;
 import run.ratchet.ri.core.internal.DefaultNodeTagAffinityProvider;
@@ -32,6 +31,7 @@ import run.ratchet.ri.core.internal.JobTimeoutHandler;
 import run.ratchet.ri.core.internal.JobWakeupService;
 import run.ratchet.ri.core.internal.OrphanRecoveryTimer;
 import run.ratchet.ri.core.internal.Poller;
+import run.ratchet.ri.core.internal.PoolRegistry;
 import run.ratchet.ri.core.internal.PostExecutionHandler;
 import run.ratchet.ri.core.internal.SingletonLeaseService;
 import run.ratchet.ri.core.internal.ThreadPoolManager;
